Fix documentation
